requirements.txt
aiobotocore==2.15.2
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.3.1
alembic==1.14.0
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==5.0.1
asyncssh==2.18.0
atpublic==5.0
attrs==24.2.0
awscli==1.36.10
billiard==4.2.1
blinker==1.9.0
boto3==1.35.36
botocore==1.35.36
cachetools==5.5.0
catboost==1.2.7
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.0
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
databricks-sdk==0.38.0
Deprecated==1.2.15
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docutils==0.16
dpath==2.2.0
dulwich==0.22.6
dvc==3.58.0
dvc-data==3.16.7
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-s3==3.2.0
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.1.0
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
funcy==2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
graphviz==0.20.3
greenlet==3.1.1
gto==1.7.2
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
iterative-telemetry==0.0.9
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kombu==5.4.2
logging==0.4.9.6
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
mlflow==2.18.0
mlflow-skinny==2.18.0
multidict==6.1.0
narwhals==1.15.2
networkx==3.4.2
numpy==1.26.4
omegaconf==2.3.0
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.0
psutil==6.1.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
pydot==3.0.3
pygit2==1.16.0
Pygments==2.18.0
pygtrie==2.5.0
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
s3fs==2024.10.0
s3transfer==0.10.4
scikit-learn==1.5.2
scipy==1.14.1
scmrepo==3.3.9
semver==3.0.2
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.36
sqlparse==0.5.2
sqltrie==0.11.1
streamlit==1.40.2
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
voluptuous==0.15.2
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.3
zc.lockfile==3.0.post1
zipp==3.21.0
